service worker: Enable nav preload network fallback test under NetworkService.
The test was flakily failing because the test server was seeing three
requests during the navigation instead of the expected one or two.

Analysis of the bug showed this was an expected possible path of the
browser cache, which happens when there are concurrent requests and one
of them is cancelled. The cache thinks it should not continue and
restarts the ongoing request, which can cause an extra network request.

This test hit the path because it issues a navigation preload request
and immediately cancels it and starts the normal navigation request.

Since developers are advised not to activate navigation preload and then
not use it,[1] we can keep this behavior.

It's unknown though why r585201 caused this flake to appear when it
didn't appear before. It's looks like it should have always been
possible.

[1] https://developers.google.com/web/updates/2017/02/navigation-preload#using_the_preloaded_response

Bug: 876911
Cq-Include-Trybots: luci.chromium.try:linux_mojo
Change-Id: I42eca116bcfffbb3c622d040343aa62980b83ee5
Reviewed-on: https://chromium-review.googlesource.com/1189514
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#586442}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ba675a6061016667486723541f9845607afb760a