Run network service browser tests on Mac 10.12 bots.
This only runs on the main waterfall, as the CQ is on 10.13. This is fine for now, given that network service is on the CQ for Windows and Linux, and there's no Mac specific networking code that this impacts.

Bug: 877996
Change-Id: I955c397920f307869c2b1c5b84a62ae620b6be9d
Reviewed-on: https://chromium-review.googlesource.com/1200308
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#589256}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69ae1491f407b33f1ee696095a1cbe12f18d22ca