Double timeout of single_process_mash_browser_tests
It may make sense to increase the number of shards instead but I am not
sufficiently knowledgable of capacity planning to do this change.

The reason for increasing the timeout is that this bot times out consistently.

TBR=rcui@chromium.org,sky@chromium.org

Bug: 874090
BUG: 847585
Change-Id: I14f9d2eccbead7ef403e346deda5442a82d5ea27
Reviewed-on: https://chromium-review.googlesource.com/1215283
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#589864}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 39254f021e7b580cf2ae546fa1f8644e912b53f9