Fix top-chrome slide visual properties synchronization
If the page exists in a maximized browser window before going to
tablet mode, the layout that results from going to tablet mode
does not change the size of the page viewport. Hence, the visual
properties of the renderer and the browser are not automatically
synchrnoized. But going to tablet mode enables the top-chrome
sliding feature (i.e. BrowserView::GetTopControlsHeight() now
returns a non-zero value).

This CL makees sure that we synchronize the visual properties
manually, so that the renderer can get the new top-controls
height whenever it changes.

BUG=867063
TEST=manual, added a browser test.

Change-Id: Ifbbc224e1de452b2582635ea8f9b4c3992a0dceb
Reviewed-on: https://chromium-review.googlesource.com/1217355
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#590525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b069c61a107c734ec68ef1d22b277f87dcc987f8