Print Preview Refresh: Add tests
Adding/modifying tests to cover some newly tweaked behaviors and a
recent regression:
(1) Destinations dialog correctly does not close if escape is pressed
while in the search box if there is any text, but does close if there
is no text.
(2) Copies setting correctly stays valid for an empty input, and the
input is reset to the default value if blurred. However, if an invalid
value is explictly entered, the setting does not stay valid or reset
to the default value when cleared.
(3) Number settings sections correctly block 'E' and '+' key events.

(2) is broken into 2 separate tests. The new interactive test for
number-settings-section validates the reset/no reset on blur behavior.
The additions to the SetCopies test verify that the setting value is
correctly set/not set for the various cases.

Bug: None
Change-Id: If7e3461c525c9cf057ac9404d5bf5da5da03e9f2
Reviewed-on: https://chromium-review.googlesource.com/1252339
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#596302}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56f8dfd096c2dda28b939332369c53923f617841