Remove VR Nougat hardcoded timeout
Removes the explicit hard timeout on VR tests on Nougat. This makes it
instead use the default of 1 hour, which is the same as the Oreo bot and
should make the tests stop timing out after the recent addition of lots
of tests.

Change-Id: Icb290a23732977b305d6d0a779c69909e401f806
Reviewed-on: https://chromium-review.googlesource.com/c/1277884
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#599336}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80bf4b17f58209b94303d15bd154d70e4d824fb9