Fuzzer owner finding to fail early on gn warning.
If a user has an args.gn which specifies an argument which isn't
actually used in the build, gn will output a warning on stdout (instead
of stderr) but otherwise continue as normal. Then gen_fuzzer_owners.py
in GetSourcesFromDeps will add a number of 'sources' which are actually
lines from the warning message. This leads to difficult to diagnose
errors later on.

Since there is no means to suppress this warning nor move it to stderr,
this change adds '--fail-on-unused-args' to the gn command so that the
owner finding will fail earlier and and more obviously.

Change-Id: I617fe3b500ada13323b77bb18c8e40b256144833
Reviewed-on: https://chromium-review.googlesource.com/c/1286731
Reviewed-by: Max Moroz <mmoroz@chromium.org>
Reviewed-by: Abhishek Arya <inferno@chromium.org>
Commit-Queue: Ben Wagner <bungeman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#600466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0bab9f8485bdb8c8dcc4460eae67824e9a54feb0