Add system ui area observer.
This will be used in an upcoming CL. We want the Picture-in-Picture
window to avoid system ui areas such as the message center and virtual
keyboard. In order to support this, this CL adds a new WM event to fire
when something changes about a 'system ui area,' which currently
includes the virtual keyboard (floating or normal) and the message
center.

Bug: b/115291749
Bug: 841886
Bug: 874545
Bug: b/112668491
Test: Added unittest
Change-Id: Ic3b0cde8800bfc98dd2ffbe0d17137a6e1a26809
Reviewed-on: https://chromium-review.googlesource.com/c/1219366
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#601832}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd0c7f53fc00ed3fd85e556f7b5b8cbe7592666e