Change the timing of Hide and losing Focus in Mash
This CL changes the order of HideImpl and giving up the focus
within DesktopWindowTreeHostMus. As is written in the comment,
this is because some code like WindowActivityWatcher assumes
the window is already invisible when the window gives up the
focus.

BUG=896080
TEST=interactive_ui_tests --enable-features=SingleProcessMash

Change-Id: I48a45bbea87c2e66d9827d2f7776b038edfca512
Reviewed-on: https://chromium-review.googlesource.com/c/1297401
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602245}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7566f8d3267fd417992de8c43da0af10bf34115f