[VK] Keyboard tray icon should toggle the virtual keyboard.
We're changing the keyboard tray action to toggle the virtual keyboard
instead of always showing the virtual keyboard.

Previously, we made it so that the VK hides when you interact with
the shelf. So we also have to exclude the keyboard tray icon from this
behaviour.

Bug: 897672
Change-Id: I019f8acf2263a18a3f9248f232e51156c0b8a349
Reviewed-on: https://chromium-review.googlesource.com/c/1298513
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602909}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3e3afec9434a3c33160d2f65a50389134c965b3