Fix all process types getting the name "Browser" in the tracelog on Android
This was causing the ProfilingProcessHostAndroidTests to fail with
network service running out of process because the network process was
getting the wrong name. Utility services only set a name on the tracelog
if there is not currently a name set:
https://cs.chromium.org/chromium/src/content/utility/utility_service_factory.cc?l=133&rcl=a2b08947596e1510f3a6fa2f65c2491eeeb4e9ec

This modifies the library loaded hooks to only use the name "Browser" if
this is actually the browser process.

Bug: 872343
Change-Id: I0af79db900a604a11aab0a80c5870a420aca920a
Reviewed-on: https://chromium-review.googlesource.com/c/1300294
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#603117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6401b7d08fb4c6cf435d189117e62a3edcef5c0