Add support for proxy bypass and retry to network servicified version of Data Reduction Proxy.
The overall design is to decide on the proxies to mark as bad, and when to restart a request as part of the DRP's URLLoaderThrottle.

This CL is complete enough to pass fallback browser_tests, however it hacks around one piece of unimplemented functionality: the DRP configuration is not yet synchronized to throttle, so it guesses it. This needs to be addressed in a follow-up.

Bug: 721403
Change-Id: If95e52aad825d6e2c3b73fd12fdfeb9e53ae6cc6
Reviewed-on: https://chromium-review.googlesource.com/c/1297238
Commit-Queue: Eric Roman <eroman@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#604687}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a00cf68a551aae618a610c27bda146d84288c206