Configure chromium.webrtc on LUCI, flip over Linux builder-tester.
This will switch over to LUCI on the existing buildbot machines.
This waterfall is a non-critical postsubmit waterfall so we
believe we can take a few days of downtime on it if necessary.
The machines in question have webcams and exotic configs, so
standing up a new waterfall is rather expensive.

The plan is
1) take linux bots out of the old master and restart
2) swarm-bootstrap current chromium.webrtc machines
2a) maybe delete old stuff off machines (?)
3) land CL to infradata that makes machines available as
   dimensions:<builder name>
4) Land this CL

Certainly, if the above fails it's going to be hard to go back
but I think it will work :)

Tbr: jbudorick@chromium.org
Bug: 877018
Change-Id: I7ddb3e8764e717b29de9c0eeb55cf6bc2ccb002a
Reviewed-on: https://chromium-review.googlesource.com/c/1288596
Reviewed-by: Patrik Höglund <phoglund@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#608316}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 254e299adaf8f1372bfeedc4ab89d48a855e4bf6