Add field trial for SequenceManager as a backend to MessageLoop.
As in cl/1378134, we are adding a field trial to the config to test the
performance of a step of the MessageLoop to SequenceManager migration. We
are testing the performance of using SequenceManager rather than
MessageLoopImpl as a backend to MessageLoop.

Bug: 891670
Change-Id: I095afb177f7e51078a6767865913f4ceab849f5f
Reviewed-on: https://chromium-review.googlesource.com/c/1379896
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Karolina Soltys <ksolt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617109}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c3ca765b34121dd01da05a15fc6567c29a7695a