Fix KioskUpdateTest.LaunchCachedOfflineEnabledAppNoNetwork with NetworkService
LaunchCachedOfflineEnabledAppNoNetwork precaches an app, and then tries
to start it with the network disabled. The app is made up of a crx file,
and webstore metadata, but the test was only waiting for the crx data
to be precached. Previously, the metadata wasn't actually precached,
and instead it was requested during the test itself, which was supposed
to be offline. This CL makes the test wait for the metadata as well as
the crx file before starting the test, so the test itself doesn't
actually have to make any network requests.

Bug: 892357
Change-Id: I0fbbaebc8185640eab67a7d3f6ce51ac500a5422
Reviewed-on: https://chromium-review.googlesource.com/c/1407737
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#622599}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b9b70828cacafc432f4762a40153dda9bde4384