Disable InterestFeedContentSuggestions for testing
A number of bots are frequently timing out when trying to start up
Chrome on the NTP. It appears that the Feed loading spinner is showing
which might be the cause (activity never enters idle). Disable
InterestFeedContentSuggestions in fieldtrial_testing_config.json to see
if that greens up the bots.

We have a number of tests that explicitly enable
InterestFeedContentSuggestions so we will still have test coverage.

BUG=942625

Change-Id: Iac4ab77e766d886da7706528b18cfee5b621376b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529738
Reviewed-by: Sky Malice <skym@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Theresa <twellington@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#641873}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec0d93d4acf77587a0284d5213a100ba3fb7e629