Traffic annotations: auto-update sheet on linux-annotator-rel
After the 'linux-annotator-rel' buildbot successfully validates the
annotations, it will write annotation contents to a Google Sheets
document.

This is not modifying the original spreadsheet (just a copy of it),
so we can test that the script works.

Bug: 878759
Change-Id: Ia4c8096b200d9025d32ed869295e0c5620beb48f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1521240
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#645368}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a129edd0e91c07837794d0ea4d90f95a1b0281d0