Make coverage tests use the layout test merger
They previously weren't using it, because they were merging profiling
data. This CL causes the layout tests on the code coverage bot to
include the default layout test merger. https://crrev.com/c/1544846
causes this to be run as part of the profiling merge.

Note that this has some copy paste of the arguments for layout tests.
Ideally this would be purely done with a mixin, but there's no way
to apply the mixin the way it would need to be. I still made this use
a mixin so that it'd be easy to change in the future.

Bug: 923711
Change-Id: I222b22cdf80425ffdd979eb6e098eef4e1695ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1562303
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#651039}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1348a3f79c8414ee1fe73c45b7b07e6b32aab26